backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
supabase==2.13.0
PyJWT==2.10.1
bcrypt==4.3.0
python-dotenv==1.1.0
